+++ src/core/launchApp.js
@@ -19,6 +19,9 @@ let executionOrder = 1; // running counter of execution order of life-cycle-hook
  *
  * **Please Note** this function uses named parameters.
  *
+ * @param {Feature[]} features the features that comprise this
+ * application.
+ *
  * @param {Aspect[]} [aspects] the set of plugable Aspects that extend
  * **feature-u**, integrating other frameworks to match your specific
  * run-time stack.<br/><br/>
@@ -26,9 +29,6 @@ let executionOrder = 1; // running counter of execution order of life-cycle-hook
  * When NO Aspects are supplied _(an atypical case)_, only the very
  * basic **feature-u** characteristics are in effect (like fassets
  * and life-cycle hooks).
- *
- * @param {Feature[]} features the features that comprise this
- * application.
  *
  * @param {registerRootAppElmCB} registerRootAppElm the callback hook
  * that registers the supplied root application element to the specific
@@ -41,15 +41,22 @@ let executionOrder = 1; // running counter of execution order of life-cycle-hook
  * 
  * Please refer to {{book.guide.detail_reactRegistration}} for more
  * details and complete examples.
-
+ *
+ * @param {showStatusCB} [showStatus] an optional callback hook that
+ * communicates a blocking "persistent" status message to the end
+ * user.
+ *
+ * Please refer to {{book.api.showStatusCB}} for more information.
+ *
  * @return {Fassets} the Fassets object used in 
  * cross-feature-communication.
  *
  * @function launchApp
  */
-export default function launchApp({aspects=[],
-                                   features,
+export default function launchApp({features,
+                                   aspects=[],
                                    registerRootAppElm,
+                                   showStatus=showStatusFallback,
                                    ...unknownArgs}={}) {
 
   logf('STARTING: - your application is now starting up');
@@ -71,6 +78,10 @@ export default function launchApp({aspects=[],
   check(registerRootAppElm,             'registerRootAppElm is required');
   check(isFunction(registerRootAppElm), 'registerRootAppElm must be a function');
 
+  // ... showStatus
+  check(showStatus,             'showStatus is required');
+  check(isFunction(showStatus), 'showStatus must be a function');
+
   // ... unrecognized named parameter
   const unknownArgKeys = Object.keys(unknownArgs);
   check(unknownArgKeys.length === 0,  `unrecognized named parameter(s): ${unknownArgKeys}`);
@@ -78,10 +89,10 @@ export default function launchApp({aspects=[],
   // ... unrecognized positional parameter
   check(arguments.length === 1,  'unrecognized positional parameters (only named parameters can be specified)');
 
-  // peform "Aspect" property validation
+  // perform "Aspect" property validation
   op.helper.validateAspectProperties(aspects);
 
-  // peform "Feature" property validation
+  // perform "Feature" property validation
   op.alch.validateFeatureContent(features, aspectMap);
 
   // prune to activeFeatures, insuring all feature.names are unique
@@ -123,10 +134,51 @@ export default function launchApp({aspects=[],
     registerRootAppElm(rootAppElm, fassets);
 //}, 0);
 
-  // apply Feature.appDidStart() life-cycle hook
-  op.flch.appDidStart(fassets, activeFeatures, aspects);
 
-  logf('COMPLETE: your application has now started');
+  // wrap the showStatus() function to prune duplicate status
+  // ... this happens INTERNALLY when monitoring the next unresolved asynchronous appInit() process
+  const priorStatus = {
+    msg: undefined,
+    err: undefined,
+  };
+  function showStatusNoDupes(msg, err) {
+
+    // no-op on duplicate back-to-back status
+    if (msg === priorStatus.msg && err === priorStatus.err) {
+      return;
+    }
+
+    // retain the status info NOW reported
+    priorStatus.msg = msg;
+    priorStatus.err = err;
+    
+    // pass-through to supplied function
+    showStatus(msg, err);
+  }
+
+  // apply Feature.appInit() life-cycle hook
+  op.flch.appInit(fassets, activeFeatures, aspects, showStatusNoDupes)
+    .then( () => {
+
+      // >>> once all async processes of feature.appInit() have completed,
+      //     continue on with our launchApp() processes
+
+      // because of the "covert" async nature of launchApp(),
+      // we report any errors via the showStatus() mechanism.
+      try {
+
+        // apply Feature.appDidStart() life-cycle hook
+        op.flch.appDidStart(fassets, activeFeatures, aspects);
+
+        logf('COMPLETE: Your application has now started');
+      }
+      catch(err) {
+        const errMsg = 'A problem was encountered in a appDidStart() life-cycle hook';
+        logf(`INCOMPLETE: Your application did NOT start ... ${errMsg}`, err);
+        showStatusNoDupes(errMsg, err);
+      }
+
+    });
 
   // expose our new App object (used in feature cross-communication)
   return fassets;
@@ -173,6 +225,44 @@ launchApp.diag = {
  * @return void
  */
 
+
+
+//***
+//*** Specification: showStatusCB
+//***
+
+/**
+ * The optional {{book.api.launchApp}} callback hook that communicates
+ * a blocking "persistent" status message to the end user.
+ *
+ * These status messages originate from the blocking that occurs in
+ * the asynchronous processes managed by the {{book.guide.appInitCB}}
+ * life-cycle-hook.
+ *
+ * By design **feature-u** has no ability to manifest messages to the
+ * end user, because this is very app-specific in styling and other
+ * heuristics.  By default (when **NO** `showStatus` parameter is
+ * supplied, **feature-u** will simply **console log** these messages.
+ *
+ * A typical manifestation of this callback is to display a running
+ * persistent SplashScreen, seeded with the supplied message. The
+ * SplashScreen should be taken down when NO message is supplied
+ * (i.e. `''`).
+ *
+ * Please refer to {{book.guide.appInitCB}} for more details and
+ * examples.
+ *
+ * @callback showStatusCB
+ * 
+ * @param {string} [msg] - the "persistent" message to display.  When
+ * NO message is supplied (i.e. `''`), **all** user notifications
+ * should be cleared _(for example, take the SplashScreen down)_.
+ * 
+ * @param {Error} [err] - an optional error to communicate to the
+ * user.
+ *
+ * @return void
+ */
 
 
 //***
@@ -251,7 +341,7 @@ op.alch.genesis = function(aspects) {
 
 op.helper.validateAspectProperties = function(aspects) {
 
-  // peform "Aspect" property validation
+  // perform "Aspect" property validation
   // NOTE 1: This is done here rather than createAspect(), because 
   //         all Aspects need to be expanded (i.e. imported) for any extendAspectProperty() to be executed)
   //         ... this will have been done by the time launchApp() is executed!!
@@ -263,7 +353,7 @@ op.helper.validateAspectProperties = function(aspects) {
     for (const propName in aspect) { // iterate over the aspects props
 
       // handle unrecognized Aspect.property
-      // ... NOTE: Aspect extended propertis have already been added to this isAspectProperty() list
+      // ... NOTE: Aspect extended properties have already been added to this isAspectProperty() list
       //           via extendAspectProperty() 
       //               executed early within the extending Aspect (i.e. in genesis())
       //               SO it is in the list at this time!!
@@ -294,7 +384,7 @@ op.alch.validateFeatureContent = function(features, aspectMap) {
   const hookSummary = aspects.map( (aspect) => `\n  Aspect.name:${aspect.name}${aspect.validateFeatureContent ? ' <-- defines: validateFeatureContent()' : ''}` );
   logf(`aspect-life-cycle-hook ... PROCESSING: Aspect.validateFeatureContent() ... ${hookCount} hooks:${hookSummary}`);
 
-  // peform "Feature" property validation
+  // perform "Feature" property validation
   // NOTE 1: This is done here rather than createFeature(), because it's the aspect's
   //         responsibility, and this is the spot where we have aspect context.
   // NOTE 2: The original source of this error is in createFeature(), 
@@ -308,7 +398,7 @@ op.alch.validateFeatureContent = function(features, aspectMap) {
       // ... built-ins are validated by createFeature()
       // ... these extra props will be processed by an Aspect
       //     ... an error condition if if no Aspect is registered to handle it
-      // ... NOTE: Aspect extended propertis have already been added to this isFeatureProperty() list
+      // ... NOTE: Aspect extended properties have already been added to this isFeatureProperty() list
       //           via extendFeatureProperty() 
       //               executed early within the extending Aspect (i.e. in genesis())
       //               SO it is in the list at this time!!
@@ -363,9 +453,9 @@ op.helper.pruneActiveFeatures = function(features) {
 };
 
 
-//*------------------------------------------------------------------------------------
+//*--------------------------------------------------------------------------------------
 //* aspect-life-cycle-hook: expandFeatureContent(fassets, activeFeatures, aspects): void
-//*------------------------------------------------------------------------------------
+//*--------------------------------------------------------------------------------------
 
 op.alch.expandFeatureContent = function(fassets, activeFeatures, aspects) {
 
@@ -392,7 +482,7 @@ op.alch.expandFeatureContent = function(fassets, activeFeatures, aspects) {
         // perform the expansion
         if (aspect.expandFeatureContent) {
           // aspect wishes to do this
-          // ... a simple process, BUT provides the hook to do more (ex: reducer tranfer of slice)
+          // ... a simple process, BUT provides the hook to do more (ex: reducer transfer of slice)
           logf(`resolving expandWithFassets() [by aspect-life-cycle-hook Aspect.name:${aspect.name}'s Aspect.expandFeatureContent()] ON Feature.name:${feature.name}'s Feature.${aspect.name} AspectContent`);
 
           errMsg = aspect.expandFeatureContent(fassets, feature);
@@ -416,9 +506,9 @@ op.alch.expandFeatureContent = function(fassets, activeFeatures, aspects) {
 
 
 
-//*--------------------------------------------------------------------------------------
+//*----------------------------------------------------------------------------------------
 //* aspect-life-cycle-hook: assembleFeatureContent(fassets, activeFeatures, aspects): void
-//*--------------------------------------------------------------------------------------
+//*----------------------------------------------------------------------------------------
 
 op.alch.assembleFeatureContent = function(fassets, activeFeatures, aspects) {
 
@@ -440,9 +530,9 @@ op.alch.assembleFeatureContent = function(fassets, activeFeatures, aspects) {
 
 
 
-//*-----------------------------------------------------------------------
+//*-------------------------------------------------------------------------
 //* aspect-life-cycle-hook: assembleAspectResources(fassets, aspects): void
-//*-----------------------------------------------------------------------
+//*-------------------------------------------------------------------------
 
 op.alch.assembleAspectResources = function(fassets, aspects) {
 
@@ -466,9 +556,9 @@ op.alch.assembleAspectResources = function(fassets, aspects) {
 
 
 
-//*----------------------------------------------------------------------
+//*------------------------------------------------------------------------
 //* helper: defineRootAppElm(fassets, activeFeatures, aspects): rootAppElm
-//*----------------------------------------------------------------------
+//*------------------------------------------------------------------------
 
 op.helper.defineRootAppElm = function(fassets, activeFeatures, aspects) {
 
@@ -495,7 +585,7 @@ op.helper.defineRootAppElm = function(fassets, activeFeatures, aspects) {
 
   // NOTE: We do NOT validate rootAppElm to insure it is non-null!
   //       - at first glance it would appear that a null rootAppElm would render NOTHING
-  //       - HOWEVER, ULTIMATLY the app code (found in the registerRootAppElm() hook) 
+  //       - HOWEVER, ULTIMATELY the app code (found in the registerRootAppElm() hook) 
   //         can display whatever it wants ... a given app may have chosen to inject it's own rootAppElm
   logf('defining-rootAppElm ... complete, rootAppElm: ', logf.elm2html(rootAppElm));
   return rootAppElm;
@@ -504,9 +594,9 @@ op.helper.defineRootAppElm = function(fassets, activeFeatures, aspects) {
 
 
 
-//*--------------------------------------------------------------------------------------
+//*----------------------------------------------------------------------------------------
 //* aspect-life-cycle-hook: initialRootAppElm(fassets, aspects, curRootAppElm): rootAppElm
-//*--------------------------------------------------------------------------------------
+//*----------------------------------------------------------------------------------------
 
 op.alch.initialRootAppElm = function(fassets, aspects, curRootAppElm) {
 
@@ -542,9 +632,9 @@ op.alch.initialRootAppElm = function(fassets, aspects, curRootAppElm) {
 
 
 
-//*-----------------------------------------------------------------------------------------
+//*-------------------------------------------------------------------------------------------
 //* feature-life-cycle-hook: appWillStart(fassets, activeFeatures, curRootAppElm): rootAppElm
-//*-----------------------------------------------------------------------------------------
+//*-------------------------------------------------------------------------------------------
 
 op.flch.appWillStart = function(fassets, activeFeatures, curRootAppElm) {
 
@@ -583,9 +673,9 @@ op.flch.appWillStart = function(fassets, activeFeatures, curRootAppElm) {
 
 
 
-//*-------------------------------------------------------------------------------------
+//*---------------------------------------------------------------------------------------
 //* aspect-life-cycle-hook: injectRootAppElm(fassets, aspects, curRootAppElm): rootAppElm
-//*-------------------------------------------------------------------------------------
+//*---------------------------------------------------------------------------------------
 
 op.alch.injectRootAppElm = function(fassets, aspects, curRootAppElm) {
 
@@ -621,9 +711,181 @@ op.alch.injectRootAppElm = function(fassets, aspects, curRootAppElm) {
 
 
 
-//*----------------------------------------------------------------------------
+//*-----------------------------------------------------------------------------------------
+//* feature-life-cycle-hook: appInit(fassets, activeFeatures, aspects, showStatus): promise
+//*-----------------------------------------------------------------------------------------
+
+op.flch.appInit = function(fassets, activeFeatures, aspects, showStatus) {
+
+  // wrap entire process
+  // ... will resolve when ALL feature.appInit() have completed!
+  return new Promise( (resolve, reject) => {
+
+    // maintain: running counter of execution order of life-cycle-hooks (unit-test related)
+    op.flch.appInit.executionOrder = executionOrder++;
+    
+    // log summary
+    const hookCount   = activeFeatures.reduce( (count, feature) => feature.appInit ? count+1 : count, 0);
+    const hookSummary = activeFeatures.map( (feature) => `\n  Feature.name:${feature.name}${feature.appInit ? ' <-- defines: appInit()' : ''}` );
+    logf(`feature-life-cycle-hook ... PROCESSING: Feature.appInit() ... ${hookCount} hooks:${hookSummary}`);
+    
+    // locate the redux app store (if any) from our aspects
+    // ... used as a convenience to pass appState/dispatch to appInit()
+    // ... we define this from the cross-aspect redux method: Aspect.getReduxStore()
+    const reduxAspect = aspects.find( aspect => aspect.getReduxStore ? true : false );
+    const [appState, dispatch] = reduxAspect 
+                               ? [reduxAspect.getReduxStore().getState(), reduxAspect.getReduxStore().dispatch]
+                               : [undefined, undefined];
+    
+    // invoke Feature.appInit() life-cycle hooks
+    // ... accomplished in AsyncInit class
+    const asyncInits = activeFeatures.reduce( (accum, feature) => {
+      if (feature.appInit) {
+        accum.push( new AsyncInit(feature, fassets, appState, dispatch, showStatus, monitorNextAsyncInit) );
+      }
+      return accum;
+    }, []);
+    
+    // monitor the FIRST incomplete AsyncInit
+    monitorNextAsyncInit();
+    
+    // actively monitor the next incomplete AsyncInit
+    function monitorNextAsyncInit() {
+    
+      // monitor the next incomplete AsyncInit
+      const nextIncompleteAsyncInit = asyncInits.find( (asyncInit) => !asyncInit.complete );
+      if (nextIncompleteAsyncInit) {
+        nextIncompleteAsyncInit.monitor();
+      }
+    
+      // otherwise (if there are NO MORE), we are done
+      else {
+        // clear status notification to user
+        showStatus(''); // ... use '', just in case we are invoking app-supplied showStatus()
+
+        // ALL feature.appInit() have completed!
+        // ... resolve our promise
+        return resolve('ALL feature.appInit() have completed!');
+      }
+    }
+
+  }); // end of ... promise
+
+};
+
+// the default showStatus callback ... merely logs status messages
+function showStatusFallback(msg='', err=null) {
+  const postMsg = `STATUS CHANGE (from appInit() life cycle hook): '${msg}'`;
+  if (err) {
+    logf.force(`${postMsg} WITH ERROR: ${err}`, err);
+  }
+  else {
+    logf.force(postMsg);
+  }
+}
+
+// our helper class to launch/monitor async processes in feature.appInit()
+class AsyncInit {
+
+  // class constructor
+  constructor(feature,       // feature KNOWN TO HAVE appInit() hook
+              fassets,
+              appState,
+              dispatch,
+              showStatusApp, // the app-specific showStatus() callback function
+              monitorNextAsyncInit) {
+
+    // carve out instance fields here
+    this.feature         = feature;       // the feature object KNOWN to have appInit() hook
+    this.showStatusApp   = showStatusApp; // the app-specific showStatus() callback function
+    this.monitored       = false;         // is process being monitored by our external tracker
+    this.complete        = false;         // is process complete (either with or without error)
+                                          // the process's current status message (defaulted to feature name)
+    this.statusMsg       = `initializing feature: ${feature.name}`;
+    this.err             = null;          // the error (if any) from this AsyncInit process
+
+    // bind our showStatus() instance method, so it can be used as a direct function
+    this.showStatus = this.showStatus.bind(this);
+
+    // invoke the feature.appInit() hook
+    logf(`feature-life-cycle-hook ... Feature.name:${feature.name} ... invoking it's defined Feature.appInit()`);
+    let optionalPromise = null;
+    try {
+      optionalPromise = feature.appInit({showStatus: this.showStatus,
+                                         fassets,
+                                         appState,
+                                         dispatch});
+    }
+    catch(err) {
+      this.err = err;
+    }
+
+    // monitor promise completion (when returned)
+    if (optionalPromise && optionalPromise.then) { // a promise was returned
+
+      optionalPromise.then( () => { // feature.appInit() finished successfully
+        logf(`AsyncInit: finished async process for feature: ${this.feature.name} - '${this.statusMsg}'`);
+
+        // consider this a completion (with lack of err)
+        this.complete = true;
+
+        // now monitor the next AsyncInit process (if any)
+        monitorNextAsyncInit();
+      })
+
+      .catch( (err) => { // feature.appInit() finished WITH error
+        logf(`AsyncInit: finished async process for feature: ${this.feature.name} - '${this.statusMsg}' WITH ERROR: ${err}`, err);
+
+        // communicate error condition to user
+        // NOTE: this retains err in self (in a "sticky" way)
+        this.showStatus(this.statusMsg, err);
+      });
+    }
+
+    // when NO promise was returned, we consider it complete "from the start"
+    // ... unless it errored out
+    else {
+      if (!this.err) {
+        this.complete = true;
+      }
+    }
+  }
+
+  // mark this AsyncInit as now being actively monitored (by our external tracking process)
+  monitor() {
+    // mark self as being actively monitored
+    this.monitored = true;
+
+    // communicate to user what we are waiting for
+    // ... and error condition (if previously errored out)
+    this.showStatus(this.statusMsg, this.err);
+  }
+
+  // our showStatus(), used by the feature.appInit() hook
+  showStatus(msg='', err=null) {
+
+    // retain the last known msg for this feature's AsyncInit
+    this.statusMsg = msg;
+
+    // retain the last known error for this feature's AsyncInit
+    // ... done so in a "sticky" way so as to NOT clobber existing error
+    if (err) { // retain latest error (when supplied) ... retaining prior err (when NOT supplied)
+      this.err = err;
+    }
+
+    // if our process is being actively monitored, pass through to the app-specific callback function
+    if (this.monitored) {
+      this.showStatusApp(msg, this.err);
+    }
+  }
+
+}
+
+
+
+//*------------------------------------------------------------------------------
 //* feature-life-cycle-hook: appDidStart(fassets, activeFeatures, aspects): void
-//*----------------------------------------------------------------------------
+//*------------------------------------------------------------------------------
 
 op.flch.appDidStart = function(fassets, activeFeatures, aspects) {