+++ docs/history.md
@@ -84,6 +84,7 @@ reference the most current release.
 
 Release           | What                                   | *When*
 ------------------|----------------------------------------|------------------
+[v2.1.0](#v2_1_0) | appInit() Life Cycle Hook              | *July 19, 2019*
 [v2.0.0](#v2_0_0) | React Hooks                            | *May 10, 2019*
 [v1.0.1](#v1_0_1) | Docs Update                            | *September 5, 2018*
 [v1.0.0](#v1_0_0) | UI Composition                         | *August 14, 2018*
@@ -93,6 +94,38 @@ Release           | What                                   | *When*
 
 <br/><br/><br/>
 ## Details:
+
+
+
+<!-- ************************************************************* -->
+<br/><br/><br/>
+<h3 id="v2_1_0" style="margin: 10px 0px; border-width: 5px 0px; padding: 5px; border-style: solid;">
+  v2.1.0 - appInit() Life Cycle Hook <i>(July 19, 2019)</i>
+</h3>
+
+[Full Docs](https://feature-u.js.org/2.1.0/)
+&bull;
+[GitHub Release](https://github.com/KevinAst/feature-u/releases/tag/v2.1.0)
+&bull;
+[GitHub Content](https://github.com/KevinAst/feature-u/tree/v2.1.0)
+&bull;
+[Diff](https://github.com/KevinAst/feature-u/compare/v2.0.0-docs...v2.1.0)
+
+**NOTE**: This release is a **non-breaking change** _(i.e. no prior
+API was impacted)_.
+
+1. **Added**: A new [`Feature.appInit()`](../2.1.0/appLifeCycle.html#appinit) Application Life Cycle
+   Hook was added, supporting blocking async initialization.
+
+2. **Added**: In support of [`Feature.appInit()`](../2.1.0/appLifeCycle.html#appinit), the
+   [`showStatus()`](../2.1.0/api.html#showStatusCB) optional callback parameter was added to
+   [`launchApp()`](../2.1.0/api.html#launchApp) allowing the application domain to
+   communicate blocking "persistent" status messages to the end user.
+
+3. **Docs**: Documentation improvements include:
+
+   - Changed the sample application reference to
+     [`eatery-nod-w`](https://github.com/KevinAst/eatery-nod-w) _(from [`eatery-nod`](https://github.com/KevinAst/eatery-nod))_.
 
 
 
@@ -176,7 +209,7 @@ significance of this hooks enhancement.
 
      This is a **must read** to get you "up to speed" quickly!
 
-   - All diagrams are now styled to be visiable in all themes _(including
+   - All diagrams are now styled to be visible in all themes _(including
      dark mode)_.
 
    - All code samples utilize a feature.js module _(previously
@@ -215,7 +248,7 @@ releases.  _A retrofit of client code is necessary (see the
    [`Cross Feature Communication`](../1.0.0/crossCommunication.html)
    has been completely re-designed to include
    [`UI Composition`](../1.0.0/crossCommunication.html#ui-composition)
-   as a core offering of **feature-u** _(thanks {{book.ext.jeffbski}} for
+   as a core offering of **feature-u** _(thanks [@jeffbski](https://github.com/jeffbski) for
    the design collaboration)_!
 
    - This refactor promotes **one solution** for all